requirements.txt
Orange3
PyQt5>=5.12.0
orangecontrib.timeseries
autogluon.timeseries